refactor: move FactoryResolverAwareTrait inside AbstractFacade
